Implemented some frontend code to interact with the backend
We can now view and add projects on the frontend! I also did some frontend styling in this commit. In addition, I also created a delete HTTP method in the Project DAO / API controller class. This method is untested (I will write a unit test for it later). Finally, I also changed the code styling so that it is easier to read.